Add code coverage
This adds OpenCover and ReportGenerator to the CI pipeline so that we
get code coverage from our unit tests.

Additionally, this change required us to refactor `src/Check.ps1` since
we could not maintain its complexity anymore. Now with this change,
`src/Check.ps1` simply calls the other scripts.

Finally, the the artefacts and tools overcrowded `src/` so now all
building and reporting is stored to `artefacts/` directory and all build
dependencies are installed in `tools/` directory. Both directories are
gitignored on purpose.