Update LICENSE.TXT to Apache 2 (#35)
This changes all the LICENSE.txt. the original license was Eclipse
Foundation since it was mandated by the legacy versioning system.
As we transferred the code to Github, we decided to change the license
to a more common and more widely used Apache 2.

The change also includes the parts related to Fraunhofer and Schneider.