Rewrite CopyLicense script
The previous script used for copying licenses (`CopyLicense.bat`) had
fairly limited logic -- it copied the license only to a pre-defined list
of subdirectories.

However, creating new projects happens often and adding license files to
the projects is error prone. Thus the more sophisticated
`CopyLicense.p1` which copies the licenses to all the subdirectories and
allows for excludes as well as includes to cover special cases.

Additionally, we add `CheckLicenses.ps1` to the continuous integration
to make sure that all the subdirectories beneath `src/` contain a
`LICENSE.txt` file (and no conflicting license files such as `LICENSE`).