Extend Opc Ua Nodeset Import
The previous OPC UA Nodeset import treated the parsing of the nodeset
in an ad-hoc manner (the nodeset was treated as a flat structure
and the submodule was parsed directly from XML without intermediate
representation).

In this change, a proper Nodeset data model is introduced
(*via* class `UaNode`). The nodeset is first parsed into this
intermediate representation, and afterwards translated into a submodel.
This makes the code more maintainable and allows for more sophisticated
features to be extracted from the Nodeset.

Further extensions on Nodeset import are expected in the near future.