Handle free and non-free IEC CDD attributes
Only some IEC CDD attributes are free and can be used without
restrictions.  Previously, we did not distinguish free and non-free
attributes.  With this patch, we only import the free attributes per
default.  If the user wants to import all attributes, they have to
activate a check box in the import dialog.  If they activate this check
box, a warning dialog mentioning the IEC CDD EULA is shown.