Fix build scripts (#69)
* Add more excludes to CheckLicenses.ps1

This patch adds the .vs, bin and packages directories to the exclusion
list in CheckLicenses.ps1 to avoid unnecessary warning messages.

Ideally, we would use the output of `git status --ignored --porcelain`
to populate the list of ignored directories.

* Set working directory in InspectCode.ps1

All build scripts use Set-Location to set the current working directory
to the location of the executed script to make sure that it can be
executed from any location.  Only InspectCode.ps1 would only work if
called from the src directory.  This patch changes InspectCode.ps1 to
also set the working directory.