Introduce caching of code inspection in CI
It takes a long time to inspect all the code from scratch (~ minutes).
This change instructs Github to keep the inspectcode caches between
the runs to accelerate the inspection.

The key of the cache was arbitrarily set to the today's date. To
invalidate the cache, simply set a different key.