Split check into style and build-test-inspect
This splits the previous check.yml workflow into two workflows so
that they can run in parallel, have different dependencies and
also give more localised feedback in case one of them fails.

The first work flow builds the solution, runs the unit tests and
inspects the code.

The second workflow checks the appearance of the code (format, width,
dead code *etc.*).

The doc-check workflow has been removed since it was quite misleading
(the workflows run on pull requests and not on pushes any more, so this
would trigger all three workflows for the doc branches).