Separate commit message checks into a workflow
This splits the commit message checks from the check styles into a
separate workflow which runs on every pull request and push.

The style checks were run only on pull request, hence ignoring the
commit messages of a branch in the pull request.

See also this issue of the checker:
https://github.com/mristin/opinionated-commit-message/issues/28