Check pull request message on edits
The default triggers for `pull_request` workflow event are:
opened, synchronize, or reopened.
(see [Actions reference](
https://docs.github.com/en/actions/reference/
events-that-trigger-workflows#pull_request)). This prevented the
checker of commit messages to re-run when we edit the title or the
body of the pull request, thus actually preventing the developer
to react to the checker's feedback.

This patch fixes the issue by explicitly including `edited` to the
triggers.