Remove references to `V20DataModel` branch (#96)
The branch `V20DataModel` used to be the development branch. It has
been merged into master and is discontinued, while all the
development should take place on the master branch.

This patch removes all the references to it to avoid future confusion.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.