Reword remarks about copyrights in CONTRIBUTING.md
This rephrasing clarifies in more detail who is the bearer of the
copyright ("legal entity" instead of "organization") and points out
that the contributors should update the main `LICENSE.txt` with
copyright notes as well as hint that there are specific plug-in
licenses.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.