Make `Build.ps1` fail when MSBuild fails (#102)
Build.ps1 previously did not check for `$LASTEXITCODE`. Though the
`ErrorActionPreference` was set to `"Stop"`, MSBuild did not write to
STDERR and hence it was not signaled as failure. This patch checks
explicitly for `$LASTEXITCODE` and throws an exception on non-zero
values.

Hence, when MSBuild fails, `Build.ps1` will fail as well, as expected.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.