Point to external screencasts in Readme (#106)
We publish the screencasts and other larger media file to a separate
external web site, namely http://www.admin-shell-io.com. This patch
points the reader to that web site and removes the references to the
`admin-shell-io` screencasts repository.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.