Rename namespace `opctest` to `AasxGenerate`
AasxGenerate is a separate program meant to demonstrate how to
automatically generate administration shells. It was originally called
`opctest`.

This refactoring renames the namespace appropriately (`AasxGenerate`).

Additionally, the dead and/or unfinished code was moved to `obsolete/`
and the source code was reformatted to fit 120 characters per line in
JetBrains Rider.