Inline `MultipleTupleDictionary.cs`
This refactoring inlines MultiTupleDictionary in the respective usage
sites as well as the custom Tuple classes.

First, the class `MultiTupleDictionary` was a misnomer (it was actually
a `MultiValueDictionary` -- the value type was not in any way
restricted).

Second, the tuples made the code quite unreadable since the definition
sites were far away from the usage sites, so the classes with
properly named fields were introduced to improve readability.