Assign all TODOs to Michael Hoffmeister
The TODO comments occurred in many different styles and had only bare
information lacking any information about the author and the
time stamp of the TODO.

This patch introduces Michael Hoffmeister as the author of all the
TODOs (since he is the main contributor) and assigns a dummy timestamp
(1970-01-01) to indicate that the time stamp of the TODO is unknown.

The workflow build-test-inspect was intentionally skipped.