Fix location restoring in scripts
The powershell scripts had a potential bug: the current location was
pushed on stacked and popped in the `finally` block after the `Main`
function has been executed. If the `Main` fiddles with the location
stack, the final location would not be properly restored.

This patch introduces a temporary variable which stores the original
location and restores it in the `finally` block.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.