Run workflows on push to master
Previously, the workflows `build-test-inspect` and `check-style` were
run only on pull requests to master, but once squashed & merged, we did
not re-run them since we needed to save computational resources (the
repository was on free plan and private, hence limited computational
hours).

This patch makes them run on every push as the repository is public now
and no limits on computational time are imposed.