Refer to the main license in our plugins
Previously, our plug-ins had all individual LICENSE.txt which were
tedious to maintain -- we had to check plug-in after plug-in which
dependencies they use.

We unify all the licenses now in the main LICENSE.txt and add only
reference to the main license file in the corresponding LICENSE.txt of
the plug-ins.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.