Test load & save chain with samples
We perform a chain of load & saves to check that the de/serialization
works.

The chain is as follows (for different extensions such as `.xml`):
- First load from AASX (package A)
- Convert package 1 to `extension` format and save as path 1
- Load from the path 1 in `extension` format (package B)
- Save package B in `extension` format to path 2

We expect the content of the two files (path 1 and path 2,
respecitvely) to be equal.

The samples are downloaded from admin-shell-io.com/samples.
The original inspiration was AasxGenerate.