Log individual checks in Check.ps1
There was no logging in Check.ps1, so a new developer was left in the
dark what tests are exactly run.

This patch adds logging so that the command of an individual failing
check can be easily copy & pasted & re-run.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.