Make histogram of issues in InspectCode.ps1
A histogram of issues in code inspection helps us set the warning
severity level in the IDE (*e.g.*, JetBrains Rider) when we need to fix
a lot of them.

For example, we can pick a single issue type, turn off all the other
warnings and focus on fixing only that particular issue type. This is
both more efficient and less error-prone as we can consider the whole
code base under that single aspect.