Remove ecl@ss and samples from releases (#157)
Restricted ecl@ss definitions and sample AASXs were included previously
in the releases.

We exclude ecl@ss definitions since we are not allowed to distribute
them. The samples are excluded as they are quite heavy, but freely
available on-line with a salient link in the Readme.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.