Set up check-release workflow (#158)
This patch sets up a workflow to build the solution for release and
package it. Finally, the release is uploaded as artifact.

Additionally, we had to change the versioning so that it is more in line
with our release policy and at-will release cycle:

* LATEST is needed for the automatic release through the check-release
  workflow.

* `Alpha` and `beta` releases indicate the level of manual testing
  preceding the release (instead of `pre` and `post` releases which
  imply fixed release deadlines).

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.