Move plug-ins in the release to plugins subfolder
The current release contained plug-ins in a separate folder of the
release so that the user had to manually copy them to
`AasxPackageExplorer\plugins` folder. After this change, the plug-ins
are automatically copied to that subfolder so that the user needs only
to unpack the archive and the package explorer is all set up.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.