Add generic import interface and IEC CDD importer
This patch adds a new project, AasxDictionaryImport, that provides a new
import dialog that can be used to import submodels into an
administration shell.  It provides a generic interface that can be
implemented for different data providers as the IEC CDD or eCl@ss, see
Model.cs.

This patch also includes a first implementation of an import data
provider for the IEC CDD in the Cdd directory.  The IEC CDD data files
should be placed in subfolders of the iec-cdd folder in the working
directory.  Alternatively, the user can chose an import folder at
runtime.

(This commit was originally written by krahlro-sick. It was impossible
to rebase the original commit since the LICENSE.txt's changed too much
in the meanwhile. mristin only re-comitted the changes and fixed the
conflicts.)