Add unit tests for IEC CDD import (#183)
This patch adds the AasxImport.Tests project that contains unit tests
for the IEC CDD import implemented in the AasxImport project.

(This commit was originally written by krahlro-sick. It was impossible
to rebase the original commit since the LICENSE.txt's changed too much
in the meanwhile. mristin only re-comitted the changes and fixed the
conflicts.)