Fix quotes issues when checking local commits
The quotes were not correctly interpreted in
`CheckPushCommitMessages.ps1`: they were not properly escaped by
powershell so the commit checker received only a part of the message.

This patch fixes the issue by properly invoking the checker script from
within `CheckPushCommitMessages.ps1`.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.