mristin/Instruct how to fix issues with commit checks
So far, the commit message checker reported the issues but the
contributors were left in the dark how to actually fix the issues.

This patch upgrades the checker to 2.2.0 (for more verbose messages)
and hints to the contributor that the pull request (and *not* the
commit message) needs to be changed.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.