Change logo in devdoc
The default logo in the documentation for developers was misleading
readers since other projects (such as GitHub) also use it. This patch
changes the logo to our organization logo.

It would be great if we could replace it with a SVG image in the future.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.