Untrack *.jpg files in Git LFS (#202)
GithHub Pages does not support Git LFS (see [this issue][1]). Since we
use *.jpg files only in the developer's documentation, the LFS tracking
is removed so that the files are properly served on GitHub Pages.

[1]: https://github.com/git-lfs/git-lfs/issues/1342

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.