Instruct developer to look closer on failed checks (#204)
This patch points the developer to have a closer look at the output of
the previous check. This is important since many users simply focus on
the error message ("something failed") and do not intuitively look up at
the previous output.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.