Make outputs of Check.ps1 visually distinguishable
Reading the output of Check.ps1 was tedious. The individual blocks were
hardly distinguishable and it was a pain to scroll between multiple runs
in the console.

This change introduces emojis, colors and timestamps so that individual
runs can be identified more easily. Moreover, it adds a prefix for each
check so that scrolling between the checks is easier.

The commit checker is put at the end of the checks since it is now
optional (we only care about the title and the description of the pull
request).

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.