Load certificate and key from Windows Certificate
Load the certificate and key from the Windows store.
PFX file and password are no longer needed.
Limitation: The current implementation does not give any hints on
which certificate to choose, as the OpenID Connect server does not
(yet?) give any hints about which (Root) CAs are supported on its
side.