Enhance Test.ps1 to run individual tests
This adds two parameters to `Test.ps1` (`-Explore` and `-Test`) so that
the developer can list all the test cases and execute individual ones or
a subset corresponding to a prefix.

This change also extends the documentation to explain how to run
individual tests with Visual Studio to offer an alternative to the
command line execution.

The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.