Refactor XML validation into a class (#214)
Previously, we only had a single static function that read the XML
schema on every XML validation. This change refactors the validation
into a separate class so that the schema does not have to be reloaded.

This is practical for validating a large number of XML samples.

We implement a table-driven test of the validation based on
the XML files directly extracted from the AASXs obtained from
http://admin-shell-io.com/samples/.

The tests revealed an inconsistency between our logic and "official"
AAS XSD schema. The problem could be fixed by "fixing" the XML and
moving the sequence of `AssetIdentificationModelRef` to the end of
XML serialization logic of an asset.