Add XML failure cases invalid by the schema
We checked so far only that the valid XML files are correctly accepted
by the XML validator. This patch introduces a suite of XML files which
are invalid according to the schema.

The error messages are stored as-are in the test resources so that we
can verify not only the acceptance/rejection logic, but the content of
the validation records as well.