Refactor command-line handling in AasxToolkit
The current parsing logic of command-line arguments from AasxToolkit
became unruly as number of commands and edge cases grew.

This patch refactors the parsing of the command-line to a separate
unit (static class `Cli`). With this new model, the command-line
parsing is specified in a declarative manner while individual
instructions are defined in separate classes.