Fix InspectCode issues missed due to caching (#244)
InspectCode caches (both local and remote, on GitHub servers) were not
invalidated when we turned on Nullable setting in
`AasxPredefinedConcepts`. This eclipsed many issues with improper
handling of nullable values.

We turn off nullables for `AasxPredefinedConcepts` again as it is too
much refactoring work to handle it at the moment.

Finally, the caching on the remote GitHub servers during the continuous
integration is also turned off to prevend similar problems in the
future.