Convert projects to SDK style (#246)
This is an update of the csproj files which previously used non-SDK
style. Non-SDK style will be eventually made obsolete so we start
updating on time.

Moreover, SDK-style csproj's allow us to use dotnet CLI and support many
custom code analyzers such as ExhaustiveMatching.Analyzer.

Finally, we had to integrate a newer version of InspectCode as a dotnet
tool (instead of nuget package, as was the case before). This newer
version of InspectCode spotted a couple of issues (including some false
positives) which we fix here as well.