Fix Check.ps1 ignoring errors
The current Check.ps1 was pretty buggy since Powershell is capricious
when it comes to executing and propagating errors from subcommand calls
(especially with `&` operator).

This patch replaces the Check.ps1 with a proper C# program that in turn
calls the individual Powershell scripts and manipulates their output to
make it more readable.

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.