Add RemoveBinAndObj.ps1 (#256)
This patch introduces a script to delete all `bin` and `obj` directories
in the `src` directory. This is sometimes necessary in case of
substantial solution changes (such as converting projects from legacy to
SDK style).

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.