Refactor and test options in Package Explorer
This patch refactors the handling of options and plugins in Package
Explorer by pushing out the singletons to the boundaries in order to
make the code easier to reason about and to test.

The reason behind this change was very opaque behavior of parsing
options from the JSON file and with parameters coupled with misdiscovery
of plugin DLLs. Finally, the patch introduces unit tests to make up for
missing documentation.