Simplify OptionsInformation in Package Explorer
There were multiple legacy options that were actually not inferred in
Package Explorer.

This patch moves them out of `OptionsInformation` class. The internal
parsing data is moved to the corresponding function, while author,
version and license information is moved to a separate class `Pref`.

Finally, two GUI tests check that the change in the logic does not break
the application.