Check for target directory in TestGui
This patch introduces a check in TestGui that the target directory
containing the binaries for the intergation tests actually exists. We
also introduce an option `-TargetDir` to manually specify the directory
(*e.g.*, if we need to test against different builds).

The workflow build-test-inspect was intentionally skipped.
The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.