Make `*.cs` and `*.xaml` headers uniform
This patch makes all the headers in the source code uniform to include
the expected copyright notice(s), the license of the software as well
as references to other open source projects used by the software.

Where there was no copyright notice previously and it was obvious that
the code was written for AASX Package Explorer we put Michael
Hoffmeister at Festo 2018-2019 as the copyright holder.

The copyright notices from third-party source codes such as Identity
Server 4 are left as-are.