Deduplicate Log singleton and LogInstance (#292)
Package Explorer uses `Log` singleton class to manage logging meant to
be reported to the user which instantiates a `LogInstance`. Previously,
the `Log` class forwarded all the calls to `LogInstance` which
duplicated a lot of code and made it hard to follow.

This patch deduplicates the calls and directs them directly to the
singleton instance.