Display repository in build-test-inspect
This patch shows the repository to the user so that we can debug fork
problems a bit faster (*e.g.*, we don't have to wait for dependencies to
be installed if there are problems with the repository name).

The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.