Send to coveralls only on pushes to master
The workflow build-test-inspect can not run on forked repositories since
the secrets are not shared (*i.e.*, only available in the main
repository and not in the fork) so that sending to coveralls fails
without the token.

This patch skips the send-to-coveralls step in all the workflows except
when we push to master.

The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.