Re-order logic for inferring Explorer's options (#296)
This patch simplifies and re-orders the logic for inferring the options
which are set for Package Explorer at the start-up:
1) Load first the default options, if available,
2) Cover the special case when a positional argument is given and
2) Parse the command-line arguments and load any additional options from
   external files on-the-fly.

The patched code makes it more explicit that multiple additional
configuration files can be passed to the Package Explorer and how the
overruling of individual options is handled.

The patch is also necessary to prevent bugs (*e.g.*, in GUI tests) as
previous code actually ignored the default options as soon as a
command-line option was given. The user actually expects the
default options to be present and only punctually overruled by the
supplied command-line arguments.