Fix naming & improve code in AdminShellPackageEnv
This patch fixes the naming in `AdminShellPackageEnv` to comply with the
underscore prefix for private members and capitalize the public members.
This makes the code more succinct and easier to track (no need to
explicitly write `this.*`).

Additionally, minor code improvements are undertaken such as case
distinction with `switch` instead of `if` *etc.*