Work on REST event message handling
* now REST updats will be consumed by
  the ContainerBase

This is very much to be thought of. This approach will mean,
that all connectors will only emit UpdateEvents but will not
update the AAS themselves. This is kind of attractive, but will
mean, that all updates to AAS must be (complicated) channeled
through the update events and might increase the complexity.

Further idea is to mark the event as "AlreadyUpdatedToAAS" and
let the above mechanism only work on demand.