Refactor AASX file repo classes
Aim to ist unify the AasxFileRepo* classes with
the PackageContainer* classes. Both are now associated
with AASX packages and the unified ones
(PackageContainerRepoItem) combine the capabilities
of detailing, WHERE a packe is stored, WHICH Ids it is
responsibele for with the actual Environment for this
package and the capability of online retrieval, update
and AAS event handling.

The goal is to have PackageCentral to index all
contemporary known AASX packages and to give
exactly one place to access, update and persist them
to all other functionalities of AASX Package Explorer.

The further goal is to provide a List of Packages to the
AASX Server, such that in long term, the AASX Server
will use PackageCentral as central data structure.