Add GUI test for IEC CDD dictionary import
This patch adds GUI tests for the IEC CDD dictionary import.  It ensures
that the import dialog can be opened using the File menu and that IEC
CDD is listed as a supported data source.  It also uses two example
data sources: an empty data source, and a data source with two very
simple example files, containing a class, two blocks and two properties.

It also introduces new utility functions to the Common class to make it
easier two write GUI tests.