Add -dict-import-dir option
This patch adds the -dict-import-dir option that can be used to set the
default search path for the Dictionary Import feature.  If the option is
not set, the current working directory is used instead.

As we no longer have to change the working directory in the
AasxDictionaryImport.Tests, we can also remove the WorkingDirectory
class.