Enhance buildForDebug to build a single project (#345)
This patch introduces the parameter `-Project` which allows the
developer to build only a single project.

This is practical when you debug the functional tests. This way the
rebuild does not have to sourt out which projects need to be rebuilt,
and that saves you time.

The workflow check-style was intentionally skipped.
The workflow check-release was intentionally skipped.