Improve dictionary import without an AASX file
In older Package Explorer versions, the main window would start with an
empty AAS tree.  In this case, the Dictionary Import function would
create a new AAS when performing the import.  In newer Package Explorer
versions, the tree always contains a "No information available"
pseudo-element.

With this patch, the Dictionary Import function creates a new AAS if is
selected while the "No information" element is selected.