Change display name for ECLASS Basic elements
If the user views an ECLASS Advanced data source, there will be
duplicate elements:  One for ECLASS Advanced, one for ECLASS Basic.
This patch adds the option to set a display name for elements in the
details view and uses this option to add the "ECLASS Basic" suffix to
the display name of ECLASS Basic classes in an ECLASS Advanced data
source.