Support local files in dictionary import
The first data provider for the dictionary import, IEC CDD, always
stored its data in directories.  Therefore, we had an "Open Directory"
button in the Dictionary Import dialog to load an additional data
source.

ECLASS however uses XML files, which could not be loaded with the
previosu solution.  As it is hard (or impossible) to show a file dialog
that supports both directories and files, we know always use files.  In
the case of IEC CDD, we remove the name of the selected file and use its
parent directory instead.