Add support for ECLASS imports (#356)
This patch adds a new data provider to the AasxImport project that
supports eCl@ss Basic XML files.  It can also read eCl@ss Advanced XML
files but does not support all elements, especially aspects.

The import is integrated into the existing submodel and submodel element
import dialogs for IEC CDD data.

* Add support for multiple application classes

* Load translations for eCl@ss data

The domain export for eCl@ss splits the data into separate files per
language.  With this patch, we load additional eCl@ss XML files for
other languages than English to merge the different translations.

* Load unit information for eCl@ss

This patch adds support for units to the eCl@ss data provider.  The data
is read from a separate UnitsML file placed in the same directory as the
main XML file.

* Document eCl@ss file naming conventions

This patch adds our assumptions about the eCl@ss patch naming
conventions to the documentation of the eCl@ss DataProvider
implementation.

* Support aspects and multiple application classes

* Change display name for ECLASS Basic elements

If the user views an ECLASS Advanced data source, there will be
duplicate elements:  One for ECLASS Advanced, one for ECLASS Basic.
This patch adds the option to set a display name for elements in the
details view and uses this option to add the "ECLASS Basic" suffix to
the display name of ECLASS Basic classes in an ECLASS Advanced data
source.

* Support local files in dictionary import

The first data provider for the dictionary import, IEC CDD, always
stored its data in directories.  Therefore, we had an "Open Directory"
button in the Dictionary Import dialog to load an additional data
source.

ECLASS however uses XML files, which could not be loaded with the
previosu solution.  As it is hard (or impossible) to show a file dialog
that supports both directories and files, we know always use files.  In
the case of IEC CDD, we remove the name of the selected file and use its
parent directory instead.

* Fix code formatting issues