Add Fetch Online option for the import dialog (#357)
This patch adds a "Fetch Online" button to the import dialog that allows
the user to retrieve data from the network, for example from the eCl@ss
REST web service.  If clicked, a FetchOnlineDialog is opened.  In this
dialog, the user can select the data provider to use and provide a query
string for the fetch.

The patch also adds support for fetching data from the network to the
DataProvider class:
- The IsFetchSupported attribute designates whether the data provider
  supports fetching data from the network.
- The FetchPrompt attribute contains a label for the query prompt to be
  used in the user interface.
- The Fetch method creates a new IDataSource instance using the data
  fetched from the network for a query string provided by the user.

This patch also contains a default implementation of these methods in
the DataProviderBase class and a simple (and incomplete) implementation
in the Eclass data provider.