Remove caching from `build-test-inspect` workflow
This patch removes the caching from the GitHub workflow as it created
problems when people added additional dependencies, but it was not
immediately obvious how to invalidate the caches.

We simply remove the caching for now instead of coming up with a more
sophisticated caching scheme due to the lack of time.