Update SemanticID IRIs of the MTP submodels
The SemanticIDs for the MTP (type and instance) submodels still used
the prefix 'http://www.admin-shell.io/mtp/v1/'. However, in the latest
version of the spec that was submitted to the IDTA, this prefix has
changed to 'https://admin-shell.io/vdi/2658/1/0/'.

This commit applies these changes for the definitions of the
PredefinedConcepts as well as for the MTPViewer and MTPControlWrapper
plugins.