Pin dependencies of AasxPluginSmdExporter (#412)
The dependencies of AasxPluginSmdExporter were simply set to latest
(following dependabot warnings from GitHub). This breaks the build as
the different DLLs are all lumped together in the same directory. Newer
versions of System.IO.Packaging and Newtonsoft.Json hence overwrite the
older ones. The executables fail at runtime as a consequence as they can
not find the older dependencies.

In the future, we should equalize all the dependencies across the
solution to avoid such conflicts.