Fix idShort generation in forms
IdShort templates used for automatic generation in form plugins sometimes carry the {0:00} format specifier. These templates weren't correctly evaluated when the element was added to a new `SubmodelElementWrapperCollection` which doesn't exist, yet. 
This was the case, for example, in the document shelf plugin when adding the very first document to a newly created submodel. The final idShort of the SMC was `Document{0:00}`.