Fix semantic id in ManufacturerDocumentation
I don't know the code base nor the definition of the documentation
submodel but when testing the plugin I noticed a behavior that I
assume is not intended. The DocumentIdDomain SMC and the DocumentId
Property inside it share the same semantic Id.
I have no clue, though, whether the proposed change would fix that
problem at all as I can't test it.