Add default valueTypes in forms
Forms didn't used to apply the allowedValueType defined in forms JSON to
properties in most cases.

This has been changed. Whenever a plugin is used to create a new
submodel, properties inside that submodel are assigned a default
valueType. When
Whenever a new submodel is created or updated through a forms plugin,
default valueTypes are applied if defined in the JSON.