Fix idShort and semanticId in DocumentShelf plugin (#389)
* Fix semantic id in ManufacturerDocumentation

I don't know the code base nor the definition of the documentation
submodel but when testing the plugin I noticed a behavior that I
assume is not intended. The DocumentIdDomain SMC and the DocumentId
Property inside it share the same semantic Id.
I have no clue, though, whether the proposed change would fix that
problem at all as I can't test it.

* Fix idShort for DocumentVersion

The idShort of DocumentVersion in the DocumentShelf plugin is now
indexed as it should be because a document can contain many versions.