Add evaluation of XML fragments based on xPath
This adds a new fragment evaluator for XML documents. The fragment needs to be a (URL-encoded) xPath expression that is then evaluated on the XML document. This also supports content=value (will return the actual content of the selected element as string) and content=paths (will return a list of xPath expressions that can be used to query children and children's children.