Add find and replace functionality (#545)
The find/ search functionality was in the code quite a while.
Actually, nobody seemed to discover it. Now, I've added the
replace functionality as well and polished the find a bit.
Finding with several options, including regex, is possible.
Replace and replace all seems to work.

For the upcoming V3.0 core works definition. The following
attributes need to be in the meta model classes:

AdminShell.MetaModelName
AdminShell.TextSearchable
AdminShell.SkipForReflection
AdminShell.SkipForSearch
Note: Fixed .csproj files because of new behaviour. See HTML comment
in AasxCsharpLib.csproj or:
https://stackoverflow.com/questions/69919664/
publish-error-found-multiple-publish-output-files-with-
the-same-relative-path

Co-authored-by: MichaHofft <michahofft2@googlemail.com>